\name{karaoke-package}
\docType{package}
\title{
Remove Vocals from a Song                    
}
\description{
Attempt to remove vocals from a stereo '.wav' recording of a song.                 
}
\details{
\tabular{ll}{
Package: \tab karaoke \cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-02-28\cr
License: GPL-3\cr
}
}
\author{
John R. Dixon  <gbm2sas@gmail.com>

Maintainer: John R. Dixon <gbm2sas@gmail.com>
}

\references{
Well known trick.
}

\keyword{ karaoke }
