\name{karaoke-package}
\alias{karaoke-package}
\docType{package}
\title{
Remove Vocals from a Song                    
}
\description{
Attempt to remove vocals from a stereo '.wav' recording of a song.                 
}
\author{
John R. Dixon  <gbm2sas@gmail.com>

Maintainer: John R. Dixon <gbm2sas@gmail.com>
}

\references{
Well known trick.
}

\keyword{ karaoke }
